update `wazo-sysconfd` events
Why:

* Moved AsteriskReloadProgressEvent from Asterisk to Sysconfd folder
